+++ src/classes/queue-getters.ts
@@ -3,7 +3,12 @@
 
 import { QueueBase } from './queue-base';
 import { Job } from './job';
-import { clientCommandMessageReg } from '../utils';
+import {
+  clientCommandMessageReg,
+  QUEUE_SCHEDULER_SUFFIX,
+  QUEUE_EVENT_SUFFIX,
+  WORKER_SUFFIX,
+} from '../utils';
 import { JobType } from '../types';
 import { Metrics } from '../interfaces';
 
@@ -298,12 +303,7 @@ export class QueueGetters<
     });
   }
 
-  /**
-   * Get worker list related to the queue.
-   *
-   * @returns - Returns an array with workers info.
-   */
-  async getWorkers(): Promise<
+  private async baseGetClients(suffix: string): Promise<
     {
       [index: string]: string;
     }[]
@@ -311,13 +311,26 @@ export class QueueGetters<
     const client = await this.client;
     const clients = await client.client('list');
     try {
-      const list = this.parseClientList(clients);
+      const list = this.parseClientList(clients, suffix);
       return list;
     } catch (err) {
       if (!clientCommandMessageReg.test((<Error>err).message)) {
         throw err;
       }
     }
+  }
+
+  /**
+   * Get worker list related to the queue.
+   *
+   * @returns - Returns an array with workers info.
+   */
+  getWorkers(): Promise<
+    {
+      [index: string]: string;
+    }[]
+  > {
+    return this.baseGetClients(WORKER_SUFFIX);
   }
 
   /**
@@ -330,17 +343,21 @@ export class QueueGetters<
       [index: string]: string;
     }[]
   > {
-    const client = await this.client;
-    const clients = await client.client('list');
-    try {
-      const list = this.parseClientList(clients, 'qs');
-      return list;
-    } catch (err) {
-      if (!clientCommandMessageReg.test((<Error>err).message)) {
-        throw err;
-      }
-    }
+    return this.baseGetClients(QUEUE_SCHEDULER_SUFFIX);
   }
+
+  /**
+   * Get queue events list related to the queue.
+   *
+   * @returns - Returns an array with queue events info.
+   */
+     async getQueueEvents(): Promise<
+     {
+       [index: string]: string;
+     }[]
+   > {
+     return this.baseGetClients(QUEUE_EVENT_SUFFIX);
+   }
 
   /**
    * Get queue metrics related to the queue.
@@ -351,7 +368,7 @@ export class QueueGetters<
    *
    * @param start - Start point of the metrics, where 0
    * is the newest point to be returned.
-   * @param end - End poinf of the metrics, where -1 is the
+   * @param end - End point of the metrics, where -1 is the
    * oldest point to be returned.
    *
    * @returns - Returns an object with queue metrics.
@@ -403,10 +420,7 @@ export class QueueGetters<
         client[key] = value;
       });
       const name = client['name'];
-      if (
-        name &&
-        name === `${this.clientName()}${suffix ? `:${suffix}` : ''}`
-      ) {
+      if (name && name === `${this.clientName()}${suffix ? `${suffix}` : ''}`) {
         client['name'] = this.name;
         clients.push(client);
       }