+++ tests/test_rate_limiter.ts
@@ -27,7 +27,7 @@ describe('Rate Limiter', function () {
   });
 
   it('should put a job into the delayed queue when limit is hit', async function () {
-    this.timeout(5000);
+    this.timeout(6000);
     const numJobs = 5;
     const worker = new Worker(queueName, async () => {}, {
       connection,
@@ -353,7 +353,7 @@ describe('Rate Limiter', function () {
       },
     });
 
-    const worker = new Worker(queueName, async job => {}, {
+    const worker = new Worker(queueName, async () => {}, {
       connection,
       limiter: {
         max: 1,
@@ -427,7 +427,7 @@ describe('Rate Limiter', function () {
       },
     });
 
-    const worker = new Worker(queueName, async job => {}, {
+    const worker = new Worker(queueName, async () => {}, {
       connection,
       limiter: {
         max: 1,
@@ -443,7 +443,7 @@ describe('Rate Limiter', function () {
         try {
           const timeDiff = Date.now() - startTime;
           // In some test envs, these timestamps can drift.
-          expect(timeDiff).to.be.gte(20);
+          expect(timeDiff).to.be.gte(15);
           expect(timeDiff).to.be.below(325);
 
           let count = 0;